Promote Guide to a incubator library
The parsing of guides in phpDocumentor will be extracted into a library
in the near future so that it could be used by others. However, at the
moment, it is still too much in flux to extract it into a separate
repository.

As preparation for that moment, and to make it easier to isolate the
tests, docs and things like examples; I have moved the guides code to an
incubator folder and actually use composer to include it as a path-based
package.

This is not a definitive version, it includes references to descriptors
and to tactician commands and handles; and I want to extract that back
into phpDocumentor itself so that the library becomes 'pure'.