Remove getNodeRendersFactory from parser
The parsing phase should only focus on creating an AST. By removing the
nodesRendererFactory from the parser we can make sure those are never used
during parsing anymore.

I introduced a new OutputFormatRenderer to have a place to store the output
specific node renderers. This allows us to simplify the output formats, and
make them more like a configuration object. The new renderer has another side
effect, that we can now set the state of the renderer based on the expected output
format. Which allows us to inject the renderer when we need to render a node.
The latter clears paths to remove the renderer from the environment.