Directive ends on complete unindent
Expecially in code blocks the first line of the directive content can have more indentation then the following ones. Overall this is true for all directives.

The general rule is that a directive ends when the indentation level of the directive declaration is found or when the file ends.

Some directives, before all the Code Block need to preserve space.

So for these empty lines at top and bottom get removed but not all spaces. Adjusted tests acordingly.