Correct title / section tests
"A blank line after a title is optional. All text blocks up to the next title of the same or higher level are included in a section (or subsection, etc.)."

The tests where the blank line before the title was missing are not correct in my eyes. At least in our files we do not have such titles.

https://docutils.sourceforge.io/docs/ref/rst/restructuredtext.html#sections