Create files with host user ids
When running docker with the php image, any command is run as root,
unless. It is possible to specify another user to run as with the --user
option though.
Note that creating the user and group on the docker image does not seem
to be required for this to work, so let us just not do it. This spares
us from creating a custom entrypoint file.