Specify cache directory for Psalm
Before 90fb333, the Psalm cache directory was presumably created in the
home directory of the root user.

After that commit, Psalm attempts to create that directory in /, and
that fails because of permission issues.

Specifying a cache directory:
- fixes that particular issue;
- more importantly, allows the cache directory to be reused.

Note that --update-baseline does not seem to benefit from the cache, so
this is really beneficial only when running vendor/bin/psalm (no
option).