[BUGFIX] Fix the DefaultNodeRenderer for array values
Up to now when the value of a node was an array (as for nodes replace and compound) there was an empty string returned in rendering. Added handling of array values (which fixes rendering of the replace directive) and added errrors in case a node that cannot be handled by the defaultnoderenderer is passed to it.

Additionally I removed the superflous new lines in rendering of inline variables