!![TASK] Replace the subparser with a starting rule in directives
This change is breaking for anyone who implemented a SubDirective with a constructor as now the parent constructor has to be called. It also fixes the bug of disappearing field lists in direcitves.

getSubParser is only used in 2 other places that can be dealed with in a follow up such that we can remove the deprecated method all together