Endorse divergence from specification for links
According to
https://www.sphinx-doc.org/en/master/usage/restructuredtext/basics.html#inline-markup,
inline markup must be separated from surrounding text by non-word characters.

Currently, the library does not seem to be checking for the spaces, and
it seems that Github's parser has the same behavior: a link between
brackets is recognized.

That behavior is nice, and since Github adopted it, it might not be that
risky. Let us add a test so that it can be relied on.