[FEATURE] Facilitate Directive Creation
A directive can now be created just by supplying a template with the name of that directive in the folder body/directive this behaviour is now similar to how we already did it with Text rules.

As an example I transfered the rubric directive to a template-only directive. Such a concept does not exist in other Markup formats anyway. It saves us both the Node and the directive.

Additionally this enables us to output unknown directives and then feature a warning (like sphinx does) instead of swallowing it and throwing an error.