!!![FEATURE] enable basic latex rendering
To be able to render different formats we needed a more
flexible system to pick the correct renderers. The main
change in this patch to introduce a OutputAwareDelegatingNodeRenderer
as the templates are using a function to render child nodes. Where
we lose the connection with the executed renderer.

This new OutputAwareDelegatingNodeRenderer will take the output format
from the context to select the correct DelegatingNodeRender that will
find the correct renderer on its turn to render the node.

To make it easier to add custom output formats using other format
renders, eg. a single page html, or pdf using the html renderers.
the compiler for DI is adjusted. This is now a more advanced compiler
pass that will autowire the configured renderers as much as possible.