[!!!][TASK] Replace the TemplatedNodeRenderer with a configurable GeneralNoderRenderer
All we need to do it really to match the nodes to their corresponding templates.

We are also able to drop the deprecated TemplateNode

When introducing a directive in an extension and that directive needs a new node it can also register a new renderer for that node.

The matching of nodes to templates can however also be done via DI.

As a proof of concept I added latex support for the confval directive.

This also frees extensions from the need to implement their own noderenderers if all that is needed is passing the node to the template (which is like 95% of nodes introduced.)

This is breaking for any extension using the TemplatedNodeRenderer.