[BUGFIX] Prevent false positives in Lexer
We had strange bugs where String like "Icon::SIZE_*" where interpreted as named refernce without having any reason to be so.

This was due to two problems:

The catchable pattern for external hyperlinks was also catching this pattern, so it was treated as a complete token.

The interpretation of the tokens in InlineLexer::gettype interpreted PART of the resulting string as named reference and marked the complete token as named_reference

I changed the catchable pattern to not catch these tokens. I am still not happy with the regex here but cannot dive into it.

I made sure that in InlineLexer::gettype() we test against the complete string as the $value passed to it is already tokenized.

This also changed the output from two functional tests that stemmed form strings with a double point beeing wrongly interpreted as hyperlinks.