Upgrade to latest league/csv
That version adds more insight on what is returned by
League\Csv::getRecords(), which is now documented as an
Iterator<array<mixed>>. Since we are using strict types, we are already
assuming that mixed will in fact always be a string or null. Let's make the
assumption clear to static analysis tools.