[TASK] Allow to load `guides.xml` when in composer mode
When building a PHAR, or using this project as a composer dependency,
the `vendor/bin/guides` binary could not be called due to a PHP error.

This is because the file `guides.xml` is referenced via
`vendor/../guides.xml`. That works in the root of the main
`render-guides` repository, but not in a composer environment.

There the correct location is `vendor/t3docs/render-guides/guides.xml`.

This PR adds a check to resolve that file in case the first try
returns FALSE for the realpath() lookup.

Generally this lookup code is a bit spaghetti-ish and we should
properly refactor this. For the time being, a hotfix like this
would allow me to continue on using `render-guides` as a dependency
in referring projects.