Introduce filesystem component
The filesystem component makes an abstraction between guides and
flysystem. This makes it easier to maintain the project when external
dependencies are introducing breaking changes. The filesystem component
is just focused on the needs of phpDocumentor and shall not be seen
as a stand alone component to be used by other projects.

(cherry picked from commit b281d2d8e911863edfbe4240294722857d135de9)

# Conflicts:
#	composer.json
#	composer.lock
#	packages/guides/composer.json