Introduce filesystem component
The filesystem component makes an abstraction between guides and
flysystem. This makes it easier to maintain the project when external
dependencies are introducing breaking changes. The filesystem component
is just focused on the needs of phpDocumentor and shall not be seen
as a stand alone component to be used by other projects.